+++ examples/http_server.js
@@ -5,4 +5,6 @@ var server = simples(80, function () {
 	console.log('HTTP server started');
 });
 
+server.mirror(8080);
+
 router(server);