+++ examples/https_server.js
@@ -2,8 +2,10 @@ var router = require('./http_router'),
 	simples = require('simples');
 
 var server = simples(443, {
-	cert: __dirname + '/ssl_certificates/server-cert.pem',
-	key: __dirname + '/ssl_certificates/server-key.pem'
+	https: {
+		cert: __dirname + '/ssl_certificates/server-cert.pem',
+		key: __dirname + '/ssl_certificates/server-key.pem'
+	}
 }, function () {
 	console.log('HTTPS server started');
 });